Split up system panels. Odd fixes.
