+++ packages/console-app/package.json
@@ -8,10 +8,10 @@
   ],
   "scripts": {
     "analyzer": "webpack --config webpack-analyzer.config.js",
-    "build:babel": "babel ./src --out-dir ./dist/es --ignore \"**/*.test.js\" --source-maps inline",
     "build": "yarn dist",
-    "dist": "yarn clean && yarn build:babel && CONFIG_FILE=config-production.yml webpack",
+    "build:babel": "babel ./src --out-dir ./dist/es --ignore \"**/*.test.js\" --source-maps inline",
     "clean": "rm -rf dist",
+    "dist": "yarn clean && yarn build:babel && CONFIG_FILE=config-production.yml webpack",
     "lint": "semistandard 'src/**/*.js'",
     "start": "CONFIG_FILE=${CONFIG_FILE:-config-testnet.yml} VERBOSE=true webpack-dev-server --mode development",
     "test": "jest --rootDir ./src --passWithNoTests --no-cache"
