+++ packages/console-app/src/version.json
@@ -1,7 +1,7 @@
 {
   "build": {
     "name":      "@dxos/console-app",
-    "buildDate": "2020-07-21T16:37:30.252Z",
-    "version":   "1.0.0-beta.21"
+    "buildDate": "2020-07-21T20:40:54.153Z",
+    "version":   "1.0.0-beta.22"
   }
 }
